refactor(kotest): merge shapes into DSL files, share the Kotlin escaper, drop dead code
Three cleanups on the kotest integration module (behavior-preserving; jvmTest green):

- Combine each emit/*Shape into its *DslFile: ChannelShape -> ChannelDslFile,
  TypeShape -> TypeDslFile, EndpointShape -> EndpointDslFile. All were single-
  consumer (EndpointShape stays referenced across the package, which is fine —
  file boundaries don't affect same-package access).

- Reuse the IR generator's identifier escaper instead of a local copy: promote
  KotlinGenerator's private String.escapeIdentifier to a public
  String.escapeKotlinIdentifier (:src:compiler:ir) and delete kotest's
  KotlinIdentifier, repointing its call sites. Removes a duplicated Kotlin
  keyword list and guarantees raw-code emission escapes names identically to
  the generator that renders the final file.

- Remove unused code:
  * ChannelCallBuilder: buildMessage/buildMessageFields/send()/send(payload)/
    sendFields + the orphaned private generatePayload (superseded by
    messageGen / send(gen)).
  * EndpointCallBuilder.buildRequest() + CallExecutor.buildRequest()
    (superseded by buildRequestGen).
  * The unreferenced typed-terminal subsystem: expecting()/collecting(),
    statusOf, the customAssertion/expectedStatuses fields,
    CallExecutor.executeEndpoint, and ContractValidator's expected-status check
    incl. the public ContractViolationKind.UnexpectedStatus. The .call() path
    (executeRequest) is untouched.
