+++ src/integration/kotest/src/commonMain/kotlin/community/flock/wirespec/integration/kotest/emit/ChannelDslFile.kt
@@ -2,6 +2,7 @@ package community.flock.wirespec.integration.kotest.emit
 
 import community.flock.wirespec.compiler.core.emit.PackageName
 import community.flock.wirespec.compiler.core.parse.ast.Channel
+import community.flock.wirespec.compiler.core.parse.ast.Reference
 import community.flock.wirespec.compiler.core.parse.ast.Refined
 import community.flock.wirespec.compiler.core.parse.ast.Type
 import community.flock.wirespec.ir.core.File
@@ -70,8 +71,6 @@ internal object ChannelDslFile {
                 getter = FunctionCall(name = Name.of("${shape.name}Generate")),
             )
 
-            // `generate.message { … }.send()` draws one payload, publishes it through the ambient
-            // channel transport, and returns it; `topic`/`key` optionally pin the destination.
             asyncFunction("send") {
                 visibility(Visibility.PUBLIC)
                 receiver(genPayload)
@@ -83,6 +82,44 @@ internal object ChannelDslFile {
                 raw("key?.let { call.key(it) }")
                 returns(rawExpr("call.send(this)"))
             }
+        }
+    }
+}
+
+internal data class ChannelShape(
+    val name: String,
+    val payloadType: String,
+    /** Payload record fields (reused from the endpoint body machinery); empty for a primitive payload. */
+    val payloadFieldShapes: List<EndpointShape.BodyFieldShape>,
+    val modelImports: List<String>,
+) {
+    companion object {
+        fun from(
+            channel: Channel,
+            types: Map<String, Type> = emptyMap(),
+            refined: Map<String, Refined> = emptyMap(),
+        ): ChannelShape {
+            val payloadRef = channel.reference
+            val payloadType = KotlinTypeMapper.map(payloadRef)
+            val payloadCustom = payloadRef as? Reference.Custom
+
+            val payloadFieldShapes = payloadCustom
+                ?.let { EndpointShape.extractBodyFields(it.value, types, refined, visited = emptySet()) }
+                ?: emptyList()
+
+            val directRefs = payloadCustom
+                ?.let { types[it.value] }
+                ?.shape?.value
+                ?.map { it.reference }
+                ?: emptyList()
+            val modelImports = EndpointShape.modelImportsFor(listOf(payloadRef) + directRefs, payloadFieldShapes, types)
+
+            return ChannelShape(
+                name = channel.identifier.value,
+                payloadType = payloadType,
+                payloadFieldShapes = payloadFieldShapes,
+                modelImports = modelImports,
+            )
         }
     }
 }
