+++ src/integration/kotest/src/commonMain/kotlin/community/flock/wirespec/integration/kotest/emit/TypeDslFile.kt
@@ -67,3 +67,27 @@ internal object TypeDslFile {
         }
     }
 }
+
+/**
+ * The Kotest-DSL view of a standalone record [Type]: its name, the [EndpointShape.BodyFieldShape]
+ * classification of its fields (reused from the endpoint body machinery), and its model imports.
+ */
+internal data class TypeShape(
+    val name: String,
+    val fields: List<EndpointShape.BodyFieldShape>,
+    val modelImports: List<String>,
+) {
+    companion object {
+        fun from(
+            type: Type,
+            types: Map<String, Type> = emptyMap(),
+            refined: Map<String, Refined> = emptyMap(),
+        ): TypeShape {
+            val name = type.identifier.value
+            val fields = EndpointShape.extractBodyFields(name, types, refined, visited = emptySet())
+            val directRefs = type.shape.value.map { it.reference }
+            val modelImports = EndpointShape.modelImportsFor(directRefs, fields, types)
+            return TypeShape(name = name, fields = fields, modelImports = modelImports)
+        }
+    }
+}
